Connections update
Touching an output and then an input in PinchConnections (Duplicate)
will now generate a rope between the two points. Todo: 1. Make it
something other than a purple material-less line. 2. Adjust parameter
"Rope resolution" (~ number of segments in rope) so that shorter
distances have not too many segments so that the line doesn't go crazy,
and not so few that there is somehow a divide by zero error.